refactor: ViolationResult 봉인 클래스를 추가 및 FoulConditionResult, InvalidMoveResult를 내부 봉인 클래스로 이동
